Refactor SmartBearUtils into BitBarUtils
