+++ bin/maze-runner
@@ -28,7 +28,6 @@ require_relative '../lib/maze/network'
 require_relative '../lib/maze/proxy'
 require_relative '../lib/maze/retry_handler'
 require_relative '../lib/maze/runner'
-require_relative '../lib/maze/smart_bear_utils'
 require_relative '../lib/maze/terminating_server'
 
 require_relative '../lib/maze/servlets/base_servlet'
